Generated by Spark: could you maybe adjust Agents -> settings?
what if there would be needed custom post request body and headers?

and also for responses - custom structure to be for showing the messages from the call inside of the tool as response?

(currently we don't support a2a protocol, mcp, but for futuere - maybe you can add there also that setting