Generated by Spark: test token auto-refresh with different time limits.  -> is really cool and working, so you don't need to include it in the gui anymore.
and also one thin, when mennu is minimized, then i get the feeling, that the top token icon is not really working as that block which is on the side menu.

i would maybe expect that the icon on the menu also reflect the time which is reamining (circle? ) or watch?

an also one thing -> when there is time remaining 0m 0s, then the gui is at side menu showing - token active - but it is not longer active.

and one more thing, if the token contains some field inside of jwt token about expiration, so when setted user will have option to show that time as remaining, but default is 15 minutes when user adds access token

wheen i google access_token, that  response for access token url call what i found on public website:

  POST
  https://api.com/auth/realms/sandbox/protocol/openid-connect/token HTTP/1.1

  Content-Type: application/x-www-form-urlencoded

  grant_type=authorization_code
  &client_id={{Client_ID}}
  &client_secret={{Secret}}
  &username={{username}}
&password={{password}}
  &grant_type=password

response:

  HTTP/1.1 200 OK
   {
    "access_token": "token",
    "expires_in": 900,
    "refresh_expires_in": 1800,
    "refresh_token": "token",
    "token_type": "bearer",
    "not-before-policy": 0,
    "session_state": "uuid4",
    "scope": ""
  }

i don't have uncoded example of access token but i think, there will be for sure some field which tels you when expired ..

also one more think, somethimes that token endpoint url is expecting CER certification validation... but i think that depends on user